requirements.txt
openai
elasticsearch
requests
pymupdf
jieba
python-dotenv
langchain_community
langchain
tiktoken
transformers>=4.51.0
torch
sentencepiece